shared_algorithm_lens logic

+++ uniswap_edpr/src/logics/shared_algorithm_lens/src/lib.rs
@@ -21,9 +21,9 @@ pub async fn calculate (targets : Vec < String >) -> LensValue {
 
   let address = v3pool_result.clone().unwrap().result.address;
   let current_tick_liquidity_str = v3pool_result.clone().unwrap().result.liquidity;
-  let current_tick_liquidity = u32::from_str_radix(current_tick_liquidity_str.trim_start_matches("0x"), 16).unwrap();
+  let current_tick_liquidity = i128::from_str_radix(current_tick_liquidity_str.trim_start_matches("0x"), 16).unwrap();
   let sqrt_ratio_x96_str = v3pool_result.clone().unwrap().result.sqrt_ratio_x96;
-  let sqrt_ratio_x96 = u32::from_str_radix(sqrt_ratio_x96_str.trim_start_matches("0x"), 16).unwrap();
+  let sqrt_ratio_x96 = u128::from_str_radix(sqrt_ratio_x96_str.trim_start_matches("0x"), 16).unwrap();
   let tick_current = v3pool_result.clone().unwrap().result.tick_current;
   let tick_spacing = v3pool_result.clone().unwrap().result.tick_spacing;
   let token0 = v3pool_result.clone().unwrap().result.token0;
